+++ crypto/des/cbc_cksm.c
@@ -59,7 +59,7 @@
 #include "des_locl.h"
 
 DES_LONG DES_cbc_cksum(const unsigned char *in, DES_cblock *output,
-		       long length, DES_key_schedule *schedule,
+		       size_t length, DES_key_schedule *schedule,
 		       const_DES_cblock *ivec)
 	{
 	register DES_LONG tout0,tout1,tin0,tin1;
