+++ crypto/des/des.h
@@ -132,20 +132,20 @@ void DES_ecb3_encrypt(const_DES_cblock *input, DES_cblock *output,
 		      DES_key_schedule *ks1,DES_key_schedule *ks2,
 		      DES_key_schedule *ks3, int enc);
 DES_LONG DES_cbc_cksum(const unsigned char *input,DES_cblock *output,
-		       long length,DES_key_schedule *schedule,
+		       size_t length,DES_key_schedule *schedule,
 		       const_DES_cblock *ivec);
 /* DES_cbc_encrypt does not update the IV!  Use DES_ncbc_encrypt instead. */
 void DES_cbc_encrypt(const unsigned char *input,unsigned char *output,
-		     long length,DES_key_schedule *schedule,DES_cblock *ivec,
+		     size_t length,DES_key_schedule *schedule,DES_cblock *ivec,
 		     int enc);
 void DES_ncbc_encrypt(const unsigned char *input,unsigned char *output,
-		      long length,DES_key_schedule *schedule,DES_cblock *ivec,
+		      size_t length,DES_key_schedule *schedule,DES_cblock *ivec,
 		      int enc);
 void DES_xcbc_encrypt(const unsigned char *input,unsigned char *output,
-		      long length,DES_key_schedule *schedule,DES_cblock *ivec,
+		      size_t length,DES_key_schedule *schedule,DES_cblock *ivec,
 		      const_DES_cblock *inw,const_DES_cblock *outw,int enc);
 void DES_cfb_encrypt(const unsigned char *in,unsigned char *out,int numbits,
-		     long length,DES_key_schedule *schedule,DES_cblock *ivec,
+		     size_t length,DES_key_schedule *schedule,DES_cblock *ivec,
 		     int enc);
 void DES_ecb_encrypt(const_DES_cblock *input,DES_cblock *output,
 		     DES_key_schedule *ks,int enc);
@@ -174,40 +174,40 @@ void DES_encrypt3(DES_LONG *data, DES_key_schedule *ks1,
 void DES_decrypt3(DES_LONG *data, DES_key_schedule *ks1,
 		  DES_key_schedule *ks2, DES_key_schedule *ks3);
 void DES_ede3_cbc_encrypt(const unsigned char *input,unsigned char *output, 
-			  long length,
+			  size_t length,
 			  DES_key_schedule *ks1,DES_key_schedule *ks2,
 			  DES_key_schedule *ks3,DES_cblock *ivec,int enc);
 void DES_ede3_cbcm_encrypt(const unsigned char *in,unsigned char *out,
-			   long length,
+			   size_t length,
 			   DES_key_schedule *ks1,DES_key_schedule *ks2,
 			   DES_key_schedule *ks3,
 			   DES_cblock *ivec1,DES_cblock *ivec2,
 			   int enc);
 void DES_ede3_cfb64_encrypt(const unsigned char *in,unsigned char *out,
-			    long length,DES_key_schedule *ks1,
+			    size_t length,DES_key_schedule *ks1,
 			    DES_key_schedule *ks2,DES_key_schedule *ks3,
 			    DES_cblock *ivec,int *num,int enc);
 void DES_ede3_ofb64_encrypt(const unsigned char *in,unsigned char *out,
-			    long length,DES_key_schedule *ks1,
+			    size_t length,DES_key_schedule *ks1,
 			    DES_key_schedule *ks2,DES_key_schedule *ks3,
 			    DES_cblock *ivec,int *num);
 
 void DES_xwhite_in2out(const_DES_cblock *DES_key,const_DES_cblock *in_white,
 		       DES_cblock *out_white);
 
-int DES_enc_read(int fd,void *buf,int len,DES_key_schedule *sched,
+int DES_enc_read(int fd,void *buf,size_t len,DES_key_schedule *sched,
 		 DES_cblock *iv);
-int DES_enc_write(int fd,const void *buf,int len,DES_key_schedule *sched,
+int DES_enc_write(int fd,const void *buf,size_t len,DES_key_schedule *sched,
 		  DES_cblock *iv);
 char *DES_fcrypt(const char *buf,const char *salt, char *ret);
 char *DES_crypt(const char *buf,const char *salt);
 void DES_ofb_encrypt(const unsigned char *in,unsigned char *out,int numbits,
-		     long length,DES_key_schedule *schedule,DES_cblock *ivec);
+		     size_t length,DES_key_schedule *schedule,DES_cblock *ivec);
 void DES_pcbc_encrypt(const unsigned char *input,unsigned char *output,
-		      long length,DES_key_schedule *schedule,DES_cblock *ivec,
+		      size_t length,DES_key_schedule *schedule,DES_cblock *ivec,
 		      int enc);
 DES_LONG DES_quad_cksum(const unsigned char *input,DES_cblock output[],
-			long length,int out_count,DES_cblock *seed);
+			size_t length,int out_count,DES_cblock *seed);
 int DES_random_key(DES_cblock *ret);
 void DES_set_odd_parity(DES_cblock *key);
 int DES_check_key_parity(const_DES_cblock *key);
@@ -221,10 +221,10 @@ int DES_set_key_checked(const_DES_cblock *key,DES_key_schedule *schedule);
 void DES_set_key_unchecked(const_DES_cblock *key,DES_key_schedule *schedule);
 void DES_string_to_key(const char *str,DES_cblock *key);
 void DES_string_to_2keys(const char *str,DES_cblock *key1,DES_cblock *key2);
-void DES_cfb64_encrypt(const unsigned char *in,unsigned char *out,long length,
+void DES_cfb64_encrypt(const unsigned char *in,unsigned char *out,size_t length,
 		       DES_key_schedule *schedule,DES_cblock *ivec,int *num,
 		       int enc);
-void DES_ofb64_encrypt(const unsigned char *in,unsigned char *out,long length,
+void DES_ofb64_encrypt(const unsigned char *in,unsigned char *out,size_t length,
 		       DES_key_schedule *schedule,DES_cblock *ivec,int *num);
 
 int DES_read_password(DES_cblock *key, const char *prompt, int verify);
