+++ crypto/des/ede_cbcm_enc.c
@@ -72,7 +72,7 @@ http://www.cs.technion.ac.il/users/wwwb/cgi-bin/tr-get.cgi/1998/CS/CS0928.ps.gz
 #include "des_locl.h"
 
 void DES_ede3_cbcm_encrypt(const unsigned char *in, unsigned char *out,
-	     long length, DES_key_schedule *ks1, DES_key_schedule *ks2,
+	     size_t length, DES_key_schedule *ks1, DES_key_schedule *ks2,
 	     DES_key_schedule *ks3, DES_cblock *ivec1, DES_cblock *ivec2,
 	     int enc)
     {
